Use typesVersions to check existence of Intl.LocalesArgument type
