+++ package.json
@@ -23,6 +23,13 @@
     ]
   },
   "types": "./index.d.ts",
+  "typesVersions": {
+    "<4.7.4": {
+      "index.d.ts": [
+        "index-lt-4.7.4.d.ts"
+      ]
+    }
+  },
   "scripts": {
     "test": "npm run build && node --no-warnings --experimental-modules --experimental-specifier-resolution=node --icu-data-dir node_modules/full-icu --loader ./test/resolve.source.mjs ./test/all.mjs",
     "test262": "TEST262=1 npm run build && node runtest262.mjs \"$@\"",
